Fix pointer casts.

+++ libclc/amdgcn-amdhsa/libspirv/workitem/get_local_size.cl
@@ -24,16 +24,16 @@ CONST_AS char * __clc_amdgcn_dispatch_ptr(void) __asm("llvm.amdgcn.dispatch.ptr"
 #endif
 
 _CLC_DEF _CLC_OVERLOAD size_t __spirv_WorkgroupSize_x() {
-    CONST_AS ushort * ptr = (CONST_AS uint *) __dispatch_ptr();
+    CONST_AS ushort * ptr = (CONST_AS ushort *) __dispatch_ptr();
     return ptr[2];
 }
 
 _CLC_DEF _CLC_OVERLOAD size_t __spirv_WorkgroupSize_y() {
-    CONST_AS ushort * ptr = (CONST_AS uint *) __dispatch_ptr();
+    CONST_AS ushort * ptr = (CONST_AS ushort *) __dispatch_ptr();
     return ptr[3];
 }
 
 _CLC_DEF _CLC_OVERLOAD size_t __spirv_WorkgroupSize_z() {
-    CONST_AS ushort * ptr = (CONST_AS uint *) __dispatch_ptr();
+    CONST_AS ushort * ptr = (CONST_AS ushort *) __dispatch_ptr();
     return ptr[4];
 }
